refactor: Edit eventbus main
