+++ feature/home/build.gradle.kts
@@ -16,6 +16,5 @@ dependencies {
     implementation(libs.google.accompanist.permissions)
     implementation(libs.gson)
     implementation(libs.lottie.compose)
-    implementation(libs.eventbus)
     //implementation(libs.androidsvg.aar)
 }
